+++ crates/onboarding/src/basics_page.rs
@@ -135,83 +135,82 @@ fn render_theme_section(tab_index: &mut isize, cx: &mut App) -> impl IntoElement
             .enumerate()
             .map(|(index, _)| {
                 let theme = &themes[index];
-            let is_selected = theme.name == current_theme_name;
-            let name = theme.name.clone();
-            let colors = cx.theme().colors();
+                let is_selected = theme.name == current_theme_name;
+                let name = theme.name.clone();
+                let colors = cx.theme().colors();
 
-            v_flex()
-                .w_full()
-                .items_center()
-                .gap_1()
-                .child(
-                    h_flex()
-                        .id(name)
-                        .relative()
-                        .w_full()
-                        .border_2()
-                        .border_color(colors.border_transparent)
-                        .rounded(ThemePreviewTile::ROOT_RADIUS)
-                        .map(|this| {
-                            if is_selected {
-                                this.border_color(colors.border_selected)
-                            } else {
-                                this.opacity(0.8).hover(|s| s.border_color(colors.border))
-                            }
-                        })
-                        .tab_index({
-                            *tab_index += 1;
-                            *tab_index - 1
-                        })
-                        .focus(|mut style| {
-                            style.border_color = Some(colors.border_focused);
-                            style
-                        })
-                        .on_click({
-                            let theme_name = theme.name.clone();
-                            let current_theme_name = current_theme_name.clone();
+                v_flex()
+                    .w_full()
+                    .items_center()
+                    .gap_1()
+                    .child(
+                        h_flex()
+                            .id(name)
+                            .relative()
+                            .w_full()
+                            .border_2()
+                            .border_color(colors.border_transparent)
+                            .rounded(ThemePreviewTile::ROOT_RADIUS)
+                            .map(|this| {
+                                if is_selected {
+                                    this.border_color(colors.border_selected)
+                                } else {
+                                    this.opacity(0.8).hover(|s| s.border_color(colors.border))
+                                }
+                            })
+                            .tab_index({
+                                *tab_index += 1;
+                                *tab_index - 1
+                            })
+                            .focus(|mut style| {
+                                style.border_color = Some(colors.border_focused);
+                                style
+                            })
+                            .on_click({
+                                let theme_name = theme.name.clone();
+                                let current_theme_name = current_theme_name.clone();
 
-                            move |_, _, cx| {
-                                write_theme_change(theme_name.clone(), theme_mode, cx);
-                                telemetry::event!(
-                                    "Welcome Theme Changed",
-                                    from = current_theme_name,
-                                    to = theme_name
-                                );
-                            }
-                        })
-                        .map(|this| {
-                            if theme_mode == ThemeAppearanceMode::System {
-                                if let Some(light_theme_name) =
-                                    LIGHT_THEMES.get(index).copied()
-                                {
-                                    let (light, dark) = (
-                                        theme_registry.get(light_theme_name).unwrap(),
-                                        theme_registry.get(DARK_THEMES[index]).unwrap(),
+                                move |_, _, cx| {
+                                    write_theme_change(theme_name.clone(), theme_mode, cx);
+                                    telemetry::event!(
+                                        "Welcome Theme Changed",
+                                        from = current_theme_name,
+                                        to = theme_name
                                     );
-                                    this.child(
-                                        ThemePreviewTile::new(light, theme_seed)
-                                            .style(ThemePreviewStyle::SideBySide(dark)),
-                                    )
+                                }
+                            })
+                            .map(|this| {
+                                if theme_mode == ThemeAppearanceMode::System {
+                                    if let Some(light_theme_name) = LIGHT_THEMES.get(index).copied()
+                                    {
+                                        let (light, dark) = (
+                                            theme_registry.get(light_theme_name).unwrap(),
+                                            theme_registry.get(DARK_THEMES[index]).unwrap(),
+                                        );
+                                        this.child(
+                                            ThemePreviewTile::new(light, theme_seed)
+                                                .style(ThemePreviewStyle::SideBySide(dark)),
+                                        )
+                                    } else {
+                                        this.child(
+                                            ThemePreviewTile::new(theme.clone(), theme_seed)
+                                                .style(ThemePreviewStyle::Bordered),
+                                        )
+                                    }
                                 } else {
                                     this.child(
                                         ThemePreviewTile::new(theme.clone(), theme_seed)
                                             .style(ThemePreviewStyle::Bordered),
                                     )
                                 }
-                            } else {
-                                this.child(
-                                    ThemePreviewTile::new(theme.clone(), theme_seed)
-                                        .style(ThemePreviewStyle::Bordered),
-                                )
-                            }
-                        }),
-                )
-                .child(
-                    Label::new(family_names[index].clone())
-                        .color(Color::Muted)
-                        .size(LabelSize::Small),
-                )
-                .into_any_element()
+                            }),
+                    )
+                    .child(
+                        Label::new(family_names[index].clone())
+                            .color(Color::Muted)
+                            .size(LabelSize::Small),
+                    )
+                    .into_any_element()
             })
             .collect()
     }